# Conflicts:
#	commands_folder/create_file.py
#	commands_folder/get_file.py
#	commands_folder/store_coins_file.py
#	global_context.py
#	main.py

+++ commands_folder/get_file.py
@@ -2,14 +2,16 @@ from global_context import *
 from commands_folder import store_coins_file
 
 async def get_command(ctx):
-    await ctx.send("You got coin, " + ctx.author.display_name + " (" + str(ctx.author.id) + ")")
-    context[0] = False
-    await context[1].delete()
-    await store_coins_file.filecheck()
-    context[2][ctx.author.id][0] += 1
-    await ctx.send("You now have "+str(context[2][ctx.author.id][0])+" coins.")
-    context[3] = 0
-    await store_coins_file.savecoins(ctx)
+    if context[0]:
+        context[6] = True
+        await ctx.send("You got coin, " + ctx.author.display_name + " (" + str(ctx.author.id) + ")")
+        context[0] = False
+        await context[1].delete()
+        context[2][ctx.author.id][0] += 1
+        await ctx.send("You now have "+str(context[2][ctx.author.id][0])+" coins.")
+        await store_coins_file.savecoins(ctx)
+    else:
+        await ctx.send("there is no coin idiot")
